requirements.txt
click==8.1.2
cycler==0.11.0
easygui==0.98.3
fonttools==4.33.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
regex==2022.4.24
six==1.16.0
tqdm==4.64.0
wordcloud==1.8.1
